+++ R/DefineClasses.r
@@ -693,7 +693,8 @@ setMethod("initialize", "lmmodel", function(.Object,Name,models){
 #' @slot CB_hist Stored historical catches in weight (historical simulations - an array with dimensions nsim, nages, nyears, nareas)
 #' @slot FM_hist Stored historical fishing mortality rate (historical simulations - an array with dimensions nsim, nages, nyears, nareas)
 #' @slot Effort Stored relative fishing effort in the projection years
-#' @slot Extra A list for additional information
+#' @slot PopAS Stored age structure of population (nsim, nMPs, nages, proyears)
+#' @slot CatchAS Stored age structure of catch (nsim, nMPs, nages, proyears)
 #'
 #' @author T. Carruthers
 #' @keywords classes
@@ -702,12 +703,12 @@ setClass("MSE", representation(Name = "character", nyears = "numeric",
   OM = "data.frame", Obs = "data.frame", B_BMSY = "array", F_FMSY = "array", 
   B = "array", SSB="array", VB="array", FM = "array", C = "array", 
   TAC = "array", SSB_hist = "array", 
-  CB_hist = "array", FM_hist = "array", Effort = "array"))
+  CB_hist = "array", FM_hist = "array", Effort = "array", PopAS = "array", CatchAS = "array"))
 
   
 setMethod("initialize", "MSE", function(.Object, Name, nyears, proyears, 
   nMPs, MPs, nsim, OM, Obs, B_BMSY, F_FMSY, B, SSB, VB, FM, C, TAC, 
-  SSB_hist, CB_hist, FM_hist, Effort = array()) {
+  SSB_hist, CB_hist, FM_hist, Effort = array(), PopAS, CatchAS) {
   .Object@Name <- Name
   .Object@nyears <- nyears
   .Object@proyears <- proyears
@@ -728,6 +729,8 @@ setMethod("initialize", "MSE", function(.Object, Name, nyears, proyears,
   .Object@CB_hist <- CB_hist
   .Object@FM_hist <- FM_hist
   .Object@Effort <- Effort
+  .Object@PopAS <- PopAS
+  .Object@CatchAS <- CatchAS
   .Object
 })